+++ FreightAccounting.WPF/AddExpenseWindow.xaml.cs
@@ -49,7 +49,7 @@ public partial class AddExpenseWindow : Window
             {
                 await expensesRepository.UpdateExpense(expenseId, new AddUpdateExpenseModel
                 {
-                    ExpensesAmount = Convert.ToInt32(txtExpensesAmount.Text.Replace(",", "")),
+                    ExpensesAmount = Convert.ToInt64(txtExpensesAmount.Text.Replace(",", "")),
                     SubmitDate = dpExpense.SelectedDate.ToDateTime(),
                 });
                 NotificationEventsManager.OnShowMessage("عملیات ویرایش با موفقیت انجام شد!", MessageTypeEnum.Success);
@@ -58,7 +58,7 @@ public partial class AddExpenseWindow : Window
             {
                 await expensesRepository.AddExpense(new AddUpdateExpenseModel
                 {
-                    ExpensesAmount = Convert.ToInt32(txtExpensesAmount.Text.Replace(",", "")),
+                    ExpensesAmount = Convert.ToInt64(txtExpensesAmount.Text.Replace(",", "")),
                     SubmitDate = dpExpense.SelectedDate.ToDateTime(),
                 });
                 NotificationEventsManager.OnShowMessage("مورد جدید با موفقیت اضافه شد!", MessageTypeEnum.Success);
